+++ Users/Bsp/Bsp_Boards/Board_Remote_Control/Board_RC.h
@@ -7,13 +7,13 @@
 extern "C"{
 #endif
 
-extern void RC_Init(UART_HandleTypeDef *huart, uint32_t *DstAddress, uint32_t *SecondMemAddress, uint32_t DataLength);
+extern void RC_Init(UART_HandleTypeDef *huart, uint8_t *DstAddress, uint8_t *SecondMemAddress, uint32_t DataLength);
 
 extern void RC_Disable(void);
 
 extern void RC_Restart(uint16_t dma_buf_num);
 
-extern void USART_RxDMA_DoubleBuffer_Init(UART_HandleTypeDef *huart, uint32_t *DstAddress, uint32_t *SecondMemAddress, uint32_t DataLength);
+extern void USART_RxDMA_DoubleBuffer_Init(UART_HandleTypeDef *huart, uint8_t *DstAddress, uint8_t *SecondMemAddress, uint32_t DataLength);
 #ifdef __cplusplus
 }
 #endif
